chore: better error handling in consent
